refactor(connector) Merge ng-redux and connector
Noticed that lots of ng-redux.ts was simply exposing things from the
store which was created from the connector.

Refactored connector so all of the ng-redux.ts functionality was moved
into it, then renamed it.

Updated tests, and cleaned up the provide factory.
