+++ package.json
@@ -4,7 +4,7 @@
   "description": "Angular 2 bindings for Redux",
   "main": "./lib/index.js",
   "scripts": {
-    "build": "rimraf ./lib; tsc",
+    "build": "rimraf ./lib; tsc; rimraf ./lib/___tests___",
     "postinstall": "typings install",
     "test": "ts-node ./node_modules/mocha/bin/_mocha  --opts ./src/___tests___/mocha.opts",
     "typings": "rimraf ./typings && typings install"
