+++ src/components/provider.ts
@@ -1,20 +1,10 @@
-import { BehaviorSubject, Observable } from 'rxjs';
-import Connector from './connector';
-import {provide, Injector, Provider, Inject, Injectable} from 'angular2/core';
-import * as Redux from 'redux';
-import {INgRedux} from './interfaces';
-import {NgRedux} from './ng-redux';
+import { provide, Provider } from 'angular2/core';
+import { Store }  from 'redux';
+import { NgRedux } from './ng-redux';
 
-export function provider<T>(store: Redux.Store<T>) {
-  const _connector = new Connector(store);
-  const factory = (): INgRedux<T> => <INgRedux<T>>Object.assign({},
-    {
-      connect: _connector.connect,
-      mapDispatchToTarget: _connector.mapDispatchToTarget
-    },
-    store);
+export function provider<T>(store: Store<T>) {
   
-  return [provide(NgRedux, { useFactory: () => new NgRedux(factory()) }),
+  return [provide(NgRedux, { useFactory: () => new NgRedux<T>(store) }),
   provide('ngRedux', {useExisting: NgRedux})];
    
 }
