+++ src/index.ts
@@ -1,2 +1,2 @@
 export {provider} from './components/provider';
-export {NgRedux} from './components/ng-redux';
+export { NgRedux }   from './components/ng-redux';
